requirements.txt
h5py==3.10.0
librosa==0.10.1
lightning==2.2.1
numpy==2.1.3
omegaconf==2.3.0
pandas==2.2.3
rich==13.9.4
scipy==1.14.1
soundfile==0.12.1
torch==2.2.1
torchaudio==2.2.1
torchmetrics==1.3.1
tqdm==4.66.4
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-joblib-launcher==1.2.0